# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: FaultWitness Control API Design
  version: 1.0.0
  description: Frozen public contract implemented by the I-0012 Control API runtime.
security:
  - oidc: [incidents:read]
paths:
  /v1/incidents:
    post:
      operationId: createIncident
      summary: Create or return an idempotent tenant-scoped Incident
      security:
        - oidc: [incidents:write]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/IncidentCreate"}
      responses:
        "201": {description: Created, content: {application/json: {schema: {$ref: "#/components/schemas/IncidentSnapshot"}}}}
        "200": {description: Existing idempotent result, content: {application/json: {schema: {$ref: "#/components/schemas/IncidentSnapshot"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
        "429": {$ref: "#/components/responses/QuotaExceeded"}
        "503": {$ref: "#/components/responses/DependencyUnavailable"}
  /v1/incidents/{incident_id}:
    parameters:
      - $ref: "#/components/parameters/IncidentId"
    get:
      operationId: getIncident
      summary: Get the current tenant-scoped Incident snapshot
      responses:
        "200": {description: Current snapshot, content: {application/json: {schema: {$ref: "#/components/schemas/IncidentSnapshot"}}}}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "503": {$ref: "#/components/responses/DependencyUnavailable"}
  /v1/incidents/{incident_id}/events:
    parameters:
      - $ref: "#/components/parameters/IncidentId"
    get:
      operationId: streamIncidentEvents
      summary: Stream truthful typed events with monotonic resume IDs
      parameters:
        - name: Last-Event-ID
          in: header
          required: false
          schema: {type: string, minLength: 1}
      responses:
        "200":
          description: Server-sent event stream; retention gaps terminate with an explicit typed event
          headers:
            Cache-Control: {schema: {type: string, const: no-cache}}
          content:
            text/event-stream:
              schema: {type: string}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "503": {$ref: "#/components/responses/DependencyUnavailable"}
  /v1/incidents/{incident_id}/approvals:
    parameters:
      - $ref: "#/components/parameters/IncidentId"
    post:
      operationId: decideApproval
      summary: Approve or reject the exact pending action digest
      security:
        - oidc: [actions:approve]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ApprovalRequest"}
      responses:
        "200": {description: Decision recorded, content: {application/json: {schema: {$ref: "#/components/schemas/ApprovalResult"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "429": {$ref: "#/components/responses/QuotaExceeded"}
        "503": {$ref: "#/components/responses/DependencyUnavailable"}
  /v1/incidents/{incident_id}/cancel:
    parameters:
      - $ref: "#/components/parameters/IncidentId"
    post:
      operationId: cancelIncident
      summary: Request cancellation at the next safe checkpoint
      security:
        - oidc: [incidents:write]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CancelRequest"}
      responses:
        "202": {description: Cancellation requested, content: {application/json: {schema: {$ref: "#/components/schemas/IncidentSnapshot"}}}}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "429": {$ref: "#/components/responses/QuotaExceeded"}
        "503": {$ref: "#/components/responses/DependencyUnavailable"}
  /v1/incidents/{incident_id}/feedback:
    parameters:
      - $ref: "#/components/parameters/IncidentId"
    post:
      operationId: recordIncidentFeedback
      summary: Append user feedback without rewriting terminal state
      security:
        - oidc: [incidents:write]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/FeedbackRequest"}
      responses:
        "202": {description: Feedback appended, content: {application/json: {schema: {$ref: "#/components/schemas/FeedbackResult"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "503": {$ref: "#/components/responses/DependencyUnavailable"}
  /v1/tools:
    get:
      operationId: listTools
      summary: List the authenticated tenant and role scoped tool catalog
      responses:
        "200": {description: Scoped tools, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/ToolSummary"}}}}}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "503": {$ref: "#/components/responses/DependencyUnavailable"}
  /v1/skills:
    get:
      operationId: listSkills
      summary: List the authenticated tenant and role scoped Skill catalog
      responses:
        "200": {description: Scoped Skills, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/SkillSummary"}}}}}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "503": {$ref: "#/components/responses/DependencyUnavailable"}
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.faultwitness.invalid/realms/faultwitness/.well-known/openid-configuration
  parameters:
    IncidentId:
      name: incident_id
      in: path
      required: true
      schema: {type: string, pattern: "^inc_[A-Za-z0-9_-]+$"}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 16, maxLength: 128}
  schemas:
    IncidentCreate:
      type: object
      additionalProperties: false
      required: [source, environment_id, service_scope, time_window, symptom_summary, mode, budget]
      properties:
        source: {type: string, minLength: 1, maxLength: 100}
        environment_id: {type: string, minLength: 1, maxLength: 128}
        service_scope: {type: array, minItems: 1, maxItems: 50, uniqueItems: true, items: {type: string}}
        time_window: {$ref: "#/components/schemas/TimeWindow"}
        symptom_summary: {type: string, minLength: 1, maxLength: 4000}
        alert_refs: {type: array, maxItems: 100, items: {type: string}}
        change_events: {type: array, maxItems: 100, items: {type: object}}
        mode: {enum: [diagnosis_only, allow_actions]}
        budget: {$ref: "#/components/schemas/AgentBudget"}
    TimeWindow:
      type: object
      additionalProperties: false
      required: [start, end]
      properties:
        start: {type: string, format: date-time}
        end: {type: string, format: date-time}
    AgentBudget:
      type: object
      additionalProperties: false
      required: [deadline, max_steps, max_model_calls, max_tokens, max_cost_usd]
      properties:
        deadline: {type: string, format: date-time}
        max_steps: {type: integer, minimum: 1}
        max_model_calls: {type: integer, minimum: 0}
        max_tokens: {type: integer, minimum: 0}
        max_cost_usd: {type: number, minimum: 0}
    ApprovalRequest:
      type: object
      additionalProperties: false
      required: [action_id, action_digest, decision, expected_state_version]
      properties:
        action_id: {type: string}
        action_digest: {type: string, pattern: "^[0-9a-f]{64}$"}
        decision: {enum: [approve, reject]}
        expected_state_version: {type: integer, minimum: 0}
        comment: {type: string, maxLength: 2000}
    CancelRequest:
      type: object
      additionalProperties: false
      required: [expected_state_version]
      properties:
        expected_state_version: {type: integer, minimum: 0}
        reason: {type: string, maxLength: 2000}
    FeedbackRequest:
      type: object
      additionalProperties: false
      required: [rating, expected_state_version]
      properties:
        rating: {type: integer, minimum: 1, maximum: 5}
        comment: {type: string, maxLength: 4000}
        expected_state_version: {type: integer, minimum: 0}
    IncidentSnapshot:
      type: object
      additionalProperties: false
      required: [incident_id, state, state_version, event_cursor]
      properties:
        incident_id: {type: string}
        state: {type: string}
        state_version: {type: integer, minimum: 0}
        event_cursor: {type: string}
        final_report_ref: {type: [string, "null"]}
    ApprovalResult:
      type: object
      required: [action_id, decision, action_digest, state_version]
      properties:
        action_id: {type: string}
        decision: {enum: [approve, reject]}
        action_digest: {type: string}
        state_version: {type: integer}
    FeedbackResult:
      type: object
      required: [feedback_id, accepted_at]
      properties:
        feedback_id: {type: string}
        accepted_at: {type: string, format: date-time}
    ToolSummary:
      type: object
      required: [tool_id, version, risk]
      properties:
        tool_id: {type: string}
        version: {type: string}
        risk: {enum: [R0, R1, R2, R3]}
    SkillSummary:
      type: object
      required: [skill_id, version]
      properties:
        skill_id: {type: string}
        version: {type: string}
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [code, message, retryable, correlation_id, details]
      properties:
        code: {type: string, pattern: "^ERR-[A-Z0-9-]+$"}
        message: {type: string, minLength: 1}
        retryable: {type: boolean}
        correlation_id: {type: string, minLength: 1}
        details: {type: object}
  responses:
    BadRequest: {description: Invalid typed request, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}}
    Unauthenticated: {description: Invalid or unavailable identity, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}}
    Forbidden: {description: Tenant scope or policy denial, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}}
    NotFound: {description: Tenant-scoped resource not found, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}}
    Conflict: {description: Idempotency digest or state version conflict, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}}
    QuotaExceeded: {description: Tenant quota or backpressure rejection, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}}
    DependencyUnavailable: {description: Required dependency unavailable, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}}